Move test to the correct test case.
`oe.common.lucene.uid.VersionsTests` is actually about testing the `_version`
field, but the test that is being moved tests `IndexVersion`.
