+++ server/src/test/java/org/elasticsearch/index/IndexVersionTests.java
@@ -196,4 +196,26 @@ public class IndexVersionTests extends ESTestCase {
             assertThat(luceneVersion, Matchers.equalTo(Lucene.parseVersionLenient(string, null)));
         }
     }
+
+    public void testLuceneVersionOnUnknownVersions() {
+        // between two known versions, should use the lucene version of the previous version
+        IndexVersion version = IndexVersionUtils.getPreviousVersion();
+        final IndexVersion nextVersion = IndexVersion.fromId(version.id() + 100);
+        if (IndexVersionUtils.allReleasedVersions().contains(nextVersion) == false) {
+            // the version is not known, we make an assumption the Lucene version stays the same
+            assertThat(version.luceneVersion(), equalTo(nextVersion.luceneVersion()));
+        } else {
+            // the version is known, the most we can assert is that the Lucene version is not earlier
+            // Version does not implement Comparable :(
+            assertTrue(nextVersion.luceneVersion().onOrAfter(version.luceneVersion()));
+        }
+
+        // too old version, major should be the oldest supported lucene version minus 1
+        version = IndexVersion.fromId(5020199);
+        assertThat(version.luceneVersion().major, equalTo(IndexVersionUtils.getFirstVersion().luceneVersion().major - 1));
+
+        // future version, should be the same version as today
+        version = IndexVersion.fromId(IndexVersion.current().id() + 100);
+        assertThat(version.luceneVersion(), equalTo(IndexVersion.current().luceneVersion()));
+    }
 }
